requirement.txt
pandas
numpy
matplotlib
tensorflow-cpu
sklearn
scipy